requirements.txt
PyYAML>=6.0
numpy>=1.26
tqdm>=4.66
rank-bm25>=0.2.2
torch>=2.2
Pillow>=10.0
PyMuPDF>=1.24
pytesseract>=0.3.10
